+++ src/Pages/BlogPost.story.jsx
@@ -2,6 +2,7 @@ import { unpack } from "@kickstartds/core/lib/storybook/helpers";
 import { Section } from "@kickstartds/base/lib/section";
 import { TeaserBox } from "@kickstartds/base/lib/teaser-box";
 import { Button } from "@kickstartds/base/lib/button";
+import { Storytelling } from "@kickstartds/content/lib/storytelling";
 import { Contact } from "@kickstartds/content/lib/contact";
 import { TextField } from "@kickstartds/form/lib/text-field";
 
@@ -10,7 +11,6 @@ import HeaderStories from "../header/Header.stories";
 import { SamplePost } from "../post/SamplePost";
 import { Default as PostAsideStory } from "../post-aside/PostAside.stories";
 import { Default as PostHeadStory } from "../post-head/PostHead.stories";
-import { Cta } from "../cta/CtaComponent";
 import { Footer } from "../footer/FooterComponent";
 import FooterStories from "../footer/Footer.stories";
 
@@ -25,19 +25,20 @@ const Page = () => (
       aside={unpack(PostAsideStory.args)}
       head={unpack(PostHeadStory.args)}
     />
-    {/* <SamplePost {...unpack(PostStory.args)} /> */}
     <Section
-      width="narrow"
+      width="full"
       mode="list"
       spaceBefore="default"
       spaceAfter="small"
+      className="l-section-style--highlight l-section--outer-width-narrow"
       gutter="large"
       headline={{
         content: "Get to know us",
         align: "center",
       }}
     >
-      <Cta
+      <Storytelling
+        backgroundColor="#eceff3"
         image={{
           order: {
             desktopImageLast: true,