+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+# v1.13.14 (Fri Mar 25 2022)
+
+#### ⚠️ Pushed to `master`
+
+- feat: fix logo sizing bug ([@fleven-kds](https://github.com/fleven-kds))
+
+#### Authors: 1
+
+- [@fleven-kds](https://github.com/fleven-kds)
+
+---
+
+# v1.13.13 (Thu Mar 24 2022)
+
+#### ⚠️ Pushed to `master`
+
+- feat: reintroduce cta-css to fix build fail ([@fleven-kds](https://github.com/fleven-kds))
+- feat: add various stories ([@fleven-kds](https://github.com/fleven-kds))
+
+#### Authors: 1
+
+- [@fleven-kds](https://github.com/fleven-kds)
+
+---
+
 # v1.13.12 (Tue Mar 22 2022)
 
 #### ⚠️ Pushed to `master`